+++ examples/json-pragmatic/src/main.rs
@@ -27,6 +27,7 @@ fn run_examples(base_path: &str) {
         ("basics/multi-objective.balance-load", None),
         ("basics/multi-objective.default", None),
         ("basics/multi-objective.maximize-value", None),
+        ("basics/priorities.value", None),
         ("basics/profiles.basic", Some(vec!["basics/profiles.basic.matrix.car", "basics/profiles.basic.matrix.truck"])),
         ("basics/relation-strict.basic", None),
         ("basics/relation-any.basic", None),