+++ vrp-pragmatic/src/constraints/order.rs
@@ -7,25 +7,25 @@ use vrp_core::models::common::ValueDimension;
 use vrp_core::models::problem::Job;
 use vrp_core::utils::compare_floats;
 
-/** Adds some extra penalty to jobs with priority bigger than 1. */
-pub struct PriorityModule {
+/** Adds some extra penalty to jobs with order bigger than 1. */
+pub struct OrderModule {
     constraints: Vec<ConstraintVariant>,
     keys: Vec<i32>,
 }
 
-impl PriorityModule {
+impl OrderModule {
     pub fn new(code: i32) -> Self {
         Self {
             constraints: vec![
-                ConstraintVariant::SoftRoute(Arc::new(PrioritySoftRouteConstraint {})),
-                ConstraintVariant::HardActivity(Arc::new(PriorityHardActivityConstraint { code })),
+                ConstraintVariant::SoftRoute(Arc::new(OrderSoftRouteConstraint {})),
+                ConstraintVariant::HardActivity(Arc::new(OrderHardActivityConstraint { code })),
             ],
             keys: vec![],
         }
     }
 }
 
-impl ConstraintModule for PriorityModule {
+impl ConstraintModule for OrderModule {
     fn accept_insertion(&self, _solution_ctx: &mut SolutionContext, _route_index: usize, _job: &Job) {}
 
     fn accept_route_state(&self, _ctx: &mut RouteContext) {}
@@ -41,26 +41,26 @@ impl ConstraintModule for PriorityModule {
     }
 }
 
-struct PrioritySoftRouteConstraint {}
+struct OrderSoftRouteConstraint {}
 
-impl SoftRouteConstraint for PrioritySoftRouteConstraint {
+impl SoftRouteConstraint for OrderSoftRouteConstraint {
     fn estimate_job(&self, solution_ctx: &SolutionContext, _: &RouteContext, job: &Job) -> f64 {
-        get_priority(job).map_or(0., |priority| {
+        get_order(job).map_or(0., |order| {
             let solution_cost = solution_ctx.get_max_cost();
             let penalty = if compare_floats(solution_cost, 0.) == Ordering::Equal { 1E9 } else { solution_cost * 2. };
 
-            (priority - 1) as f64 * penalty
+            (order - 1) as f64 * penalty
         })
     }
 }
 
-struct PriorityHardActivityConstraint {
+struct OrderHardActivityConstraint {
     code: i32,
 }
 
-impl PriorityHardActivityConstraint {
-    fn check_priorities(&self, first: &Job, second: &Job, stopped: bool) -> Option<ActivityConstraintViolation> {
-        let result = get_priority(first).unwrap_or(1) <= get_priority(second).unwrap_or(1);
+impl OrderHardActivityConstraint {
+    fn check_order(&self, first: &Job, second: &Job, stopped: bool) -> Option<ActivityConstraintViolation> {
+        let result = get_order(first).unwrap_or(1) <= get_order(second).unwrap_or(1);
 
         if result {
             None
@@ -70,7 +70,7 @@ impl PriorityHardActivityConstraint {
     }
 }
 
-impl HardActivityConstraint for PriorityHardActivityConstraint {
+impl HardActivityConstraint for OrderHardActivityConstraint {
     fn evaluate_activity(
         &self,
         _: &RouteContext,
@@ -85,16 +85,16 @@ impl HardActivityConstraint for PriorityHardActivityConstraint {
         match (prev, target, next) {
             (_, None, _) => None,
             (None, Some(_), None) => None,
-            (Some(prev), Some(target), _) => self.check_priorities(&prev, &target, false),
-            (_, Some(target), Some(next)) => self.check_priorities(&target, &next, true),
+            (Some(prev), Some(target), _) => self.check_order(&prev, &target, false),
+            (_, Some(target), Some(next)) => self.check_order(&target, &next, true),
         }
     }
 }
 
-fn get_priority(job: &Job) -> Option<i32> {
+fn get_order(job: &Job) -> Option<i32> {
     match job {
-        Job::Single(job) => job.dimens.get_value::<i32>("priority"),
-        Job::Multi(job) => job.dimens.get_value::<i32>("priority"),
+        Job::Single(job) => job.dimens.get_value::<i32>("order"),
+        Job::Multi(job) => job.dimens.get_value::<i32>("order"),
     }
     .cloned()
 }